pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptterfly"
version = "0.1.0"
description = "Local CLI tool for managing and optimizing prompts with simple versioning"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rich>=13.0",
    "litellm>=2.0",
    "dspy>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
promptterfly = "promptterfly.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
